+++ Modular Gameplay Overhaul/Modules/Weapons/Patchers/Infinity/EventAnswerDialoguePatcher.cs
@@ -39,7 +39,7 @@ internal sealed class EventAnswerDialoguePatcher : HarmonyPatcher
             case 8185291 when answerChoice == 1:
 
                 player.Increment(DataKeys.ProvenHonor);
-                WeaponsModule.State.Quest?.UpdateVirtueProgress(Virtue.Honor);
+                WeaponsModule.State.VirtuesQuest?.UpdateVirtueProgress(Virtue.Honor);
                 return;
 
         // COMPASSION //
@@ -60,7 +60,7 @@ internal sealed class EventAnswerDialoguePatcher : HarmonyPatcher
             case 8185294 when answerChoice == 0:
 
                 player.Increment(DataKeys.ProvenCompassion);
-                WeaponsModule.State.Quest?.UpdateVirtueProgress(Virtue.Compassion);
+                WeaponsModule.State.VirtuesQuest?.UpdateVirtueProgress(Virtue.Compassion);
                 return;
 
             // Pam 9 hearts | Location: Trailer_Big
@@ -75,7 +75,7 @@ internal sealed class EventAnswerDialoguePatcher : HarmonyPatcher
             case 27 when answerChoice == 2:
 
                 player.Increment(DataKeys.ProvenWisdom);
-                WeaponsModule.State.Quest?.UpdateVirtueProgress(Virtue.Wisdom);
+                WeaponsModule.State.VirtuesQuest?.UpdateVirtueProgress(Virtue.Wisdom);
                 return;
         }
     }