+++ Modular Gameplay Overhaul/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file contains a TL;DR of current version changes and hotfixes from across all modules. For the complete changelog, please refer to the individual changelogs of each module, linked [below](#detailed-change-logs).
 
+## Patch 2.2.7 Highlights <sup><sub><sup>[🔼](#margo-change-logs)</sup></sub></sup>
+
+* [WPNZ]: Fixed Blade of Ruin duplicating itself when stored in a chest.
+* [WPNZ]: You can now see and change all Stabbing Swords using a single config setting.
+* [WPNZ]: Touched up some textures.
+
 ## Patch 2.2.6 Highlights <sup><sub><sup>[🔼](#margo-change-logs)</sup></sub></sup>
 
 * Fixed a bug preventing the config menu from reloading correctly.